+++ src/application/structs/LocationData.h
@@ -7,13 +7,14 @@ enum LocationType
 {
     LocationUpgrade,
     LocationLevel,
+    LocationSubLevel,
     LocationFieldEmblem,
     LocationUnknown
 };
 
 struct LocationData
 {
-    LocationData() : eventFlag(static_cast<StoryFlags>(0x00)),
+    LocationData() : eventFlag(static_cast<StoryFlags>(0x00)), character(0), level(0), mission(0), emblemId(0),
                      type(LocationUnknown),
                      checked(false),
                      displayName(std::string("Unknown")),
@@ -41,17 +42,23 @@ struct LocationData
         return LocationData(eventFlag, -1, -1, -1, -1, LocationUpgrade, false, displayName, originalItemId);
     }
 
-    static LocationData LevelLocation(int character, int level, int mission)
+    static LocationData LevelLocation(const int character, const int level, const int mission)
     {
         return LocationData(static_cast<StoryFlags>(0x0), character, level, mission, -1, LocationLevel, false,
                             std::to_string(character) + std::to_string(level) + std::to_string(mission), -1);
     }
 
-    static LocationData FieldEmblemLocation(int emblemId, const std::string& displayName)
+    static LocationData SubLevelLocation(const int level, const int mission)
     {
-        return LocationData(static_cast<StoryFlags>(0x0), -1, -1, -1, emblemId, LocationFieldEmblem, false, displayName, -1);
+        return LocationData(static_cast<StoryFlags>(0x0), -1, level, mission, -1, LocationSubLevel, false,
+                            std::to_string(level) + std::to_string(mission), -1);
     }
 
+    static LocationData FieldEmblemLocation(const int emblemId, const std::string& displayName)
+    {
+        return LocationData(static_cast<StoryFlags>(0x0), -1, -1, -1, emblemId, LocationFieldEmblem, false, displayName,
+                            -1);
+    }
 
 
     StoryFlags eventFlag;