+++ src/output/locationRepository/LocationRepository.cpp
@@ -169,27 +169,12 @@ LocationRepository::LocationRepository()
         {6302, LocationData::LevelLocation(Characters_Big, LevelIDs_HotShelter, MISSION_C)},
 
         //Sub levels
+        {15, LocationData::SubLevelLocation(LevelIDs_TwinkleCircuit, MISSION_B)},
+        {16, LocationData::SubLevelLocation(LevelIDs_TwinkleCircuit, MISSION_A)},
 
-        {15, LocationData::LevelLocation(Characters_Sonic, LevelIDs_TwinkleCircuit, MISSION_B)},
-        {15, LocationData::LevelLocation(Characters_Tails, LevelIDs_TwinkleCircuit, MISSION_B)},
-        {15, LocationData::LevelLocation(Characters_Knuckles, LevelIDs_TwinkleCircuit, MISSION_B)},
-        {15, LocationData::LevelLocation(Characters_Amy, LevelIDs_TwinkleCircuit, MISSION_B)},
-        {15, LocationData::LevelLocation(Characters_Big, LevelIDs_TwinkleCircuit, MISSION_B)},
-        {15, LocationData::LevelLocation(Characters_Gamma, LevelIDs_TwinkleCircuit, MISSION_B)},
+        {25, LocationData::SubLevelLocation(LevelIDs_SandHill, MISSION_B)},
+        {26, LocationData::SubLevelLocation(LevelIDs_SandHill, MISSION_A)},
 
-        {16, LocationData::LevelLocation(Characters_Sonic, LevelIDs_TwinkleCircuit, MISSION_A)},
-        {16, LocationData::LevelLocation(Characters_Tails, LevelIDs_TwinkleCircuit, MISSION_A)},
-        {16, LocationData::LevelLocation(Characters_Knuckles, LevelIDs_TwinkleCircuit, MISSION_A)},
-        {16, LocationData::LevelLocation(Characters_Amy, LevelIDs_TwinkleCircuit, MISSION_A)},
-        {16, LocationData::LevelLocation(Characters_Big, LevelIDs_TwinkleCircuit, MISSION_A)},
-        {16, LocationData::LevelLocation(Characters_Gamma, LevelIDs_TwinkleCircuit, MISSION_A)},
-
-        {25, LocationData::LevelLocation(Characters_Sonic, LevelIDs_SandHill, MISSION_B)},
-        {25, LocationData::LevelLocation(Characters_Tails, LevelIDs_SandHill, MISSION_B)},
-
-        {26, LocationData::LevelLocation(Characters_Sonic, LevelIDs_SandHill, MISSION_B)},
-        {26, LocationData::LevelLocation(Characters_Tails, LevelIDs_SandHill, MISSION_A)},
-        
         // Field Emblems
         {10, LocationData::FieldEmblemLocation(119, "Station Emblem")},
         {11, LocationData::FieldEmblemLocation(121, "Burger Shop Emblem")},